feat(sql): support array literal
The `ARRAY` literal has been added, allowing to create one-dimensional typed
arrays. Supported element types: `BOOLEAN`, `DATETIME`, `DECIMAL`, `DOUBLE`,
`INTEGER`, `JSON`, `TEXT`, `UUID`.  The type is inferred via
`analyze_homogeneous_exprs` using the target column's type. All type-related
logic is implemented in Picodata; Tarantool stores untyped arrays. Supported
pgwire array encoding in binary and text format. Added `_pico_array_cast`
builtin function for `::T[]` syntax.

+++ sql-planner/src/frontend/sql/ast/ir_populator/params.rs
@@ -100,7 +100,7 @@ pub(in crate::frontend::sql) fn parse_param<M: Metadata>(
         )));
     }
 
-    let ty = param_types.get((index - 1) as usize);
+    let ty = param_types.get(index - 1);
     let ty = ty.cloned().unwrap_or(DerivedType::unknown());
     Ok(plan.add_param(index.try_into().expect("invalid parameter idnex"), ty))
 }
